Moved all functions from main.spatial_utils to spatial.utils
